+++ DBReport/My Project/Settings.Designer.vb
@@ -521,6 +521,18 @@ Namespace My
                 Me("DisplayStandardSqlType") = value
             End Set
         End Property
+        
+        <Global.System.Configuration.UserScopedSettingAttribute(),  _
+         Global.System.Diagnostics.DebuggerNonUserCodeAttribute(),  _
+         Global.System.Configuration.DefaultSettingValueAttribute("True")>  _
+        Public Property DisplayFieldLength() As Boolean
+            Get
+                Return CType(Me("DisplayFieldLength"),Boolean)
+            End Get
+            Set
+                Me("DisplayFieldLength") = value
+            End Set
+        End Property
     End Class
 End Namespace