+++ DBReport/Src/_modConst.vb
@@ -4,7 +4,7 @@
 
 Public Module _modConst
 
-    Public Const sAppDate$ = "19/07/2025"
+    Public Const sAppDate$ = "25/07/2025"
 
 #If DEBUG Then
     Public Const bDebug As Boolean = True